Update homedevice.ino
This code is functional with button and buzzer. Code is sloppy as a result of panick coding and frankensteining during the time crunch. ESP8266 has non functional pin (D0- GPIO16) because I accidently short circuited a portion of the board early today. That did come back to haunt me lmao.